Add seeder and factory for the seeder
The seeder is now working properly, default password is 'secret'
and email addresses are <yourname>@hanze.nl. This also fixes a
mistake that was made in hotels_routes table. It now has a propper
primary key. It's necessary to delete all tables and run migrate again.

This is because migration can not edit collumn types.